pyproject.toml
[project]
name = "restframework"
version = "0.1.0"
description = ""
authors = [
    {name = "Rostyslav Bysko",email = "brn.h74@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "djoser (>=2.3.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
